UI laser overhaul
