+++ Slider/Assets/Scripts/Map/Desert/DinoLaser.cs
@@ -5,6 +5,8 @@ public class DinoLaser : MonoBehaviour
     [SerializeField] private MagiLaser laser;
     [SerializeField] private SpriteRenderer spriteRenderer;
 
+    public UILaserManager uILaserManager;
+
     public void EnableSpriteRenderer(bool on)
     {
         spriteRenderer.enabled = on;
@@ -18,6 +20,6 @@ public class DinoLaser : MonoBehaviour
     public void EnableLaser(bool on)
     {
         laser.SetEnabled(on);
-        ArtifactTBPluginLaser.UpdateSpritesFromSource();
+        uILaserManager.UpdateSpritesFromSource();
     }
 }
