+++ Slider/Assets/Scripts/UI/Magitech.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7d14fd7328078fc44a7259333ac552c4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
