+++ Slider/Assets/Scripts/UI/Magitech/UILaserManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f27d57aac0fdf1340a334b129101d194
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
